+++ bulbs/api/metadata.py
@@ -1,13 +1,16 @@
 from rest_framework.metadata import SimpleMetadata
 
 from bulbs.infographics.metadata import InfographicMetadata
-from bulbs.infographics.serializers import InfographicSerializer
+from bulbs.infographics.utils import get_infographics_serializer
+
+
+INFOGRAPHICS_SERIALIZER = get_infographics_serializer()
 
 
 class PolymorphicContentMetadata(SimpleMetadata):
 
     serializer_lookup = {
-        InfographicSerializer: InfographicMetadata()
+        INFOGRAPHICS_SERIALIZER: InfographicMetadata()
     }
 
     def determine_metadata(self, request, view):